[FEAT] 제품기능 구현 #9

+++ src/main/java/com/umc/lifesharing/user/dto/UserResponseDTO.java
@@ -17,7 +17,7 @@ public class UserResponseDTO {
     public static class ProductPreviewDTO{
         String name;
         String location;
-        Float score;
+        Integer score;
         Integer score_count;
         Integer deposit;
         Integer day_price;
